urlshorteningapp/requirements.txt
Flask==3.1.1
python-dotenv>=0.21.0
requests>=2.31.0
mysql-connector-python>=8.1.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-RESTful>=0.3.9
hashids